+++ composer/autoload_namespaces.php
@@ -11,7 +11,7 @@ return array(
     'TextCat' => array($vendorDir . '/wikimedia/textcat/src'),
     'Pimple' => array($vendorDir . '/pimple/pimple/src'),
     'Net' => array($vendorDir . '/pear/net_smtp', $vendorDir . '/pear/net_socket'),
-    'Mail' => array($vendorDir . '/pear/mail_mime', $vendorDir . '/pear/mail'),
+    'Mail' => array($vendorDir . '/pear/mail', $vendorDir . '/pear/mail_mime'),
     'Liuggio' => array($vendorDir . '/liuggio/statsd-php-client/src'),
     'Less' => array($vendorDir . '/wikimedia/less.php/lib'),
     'DataValues\\' => array($vendorDir . '/data-values/number/src', $vendorDir . '/data-values/time/src'),