+++ wikimedia/minify/src/MappingsGenerator.php
@@ -94,10 +94,10 @@ class MappingsGenerator {
 					"\n"
 				) + $this->curSourceOffset + 1;
 			$this->curSourceLine += $lineCount;
-			$this->curSourceColumn = self::getJsLength(
+			$this->curSourceColumn = Utils::getJsLength(
 				substr( $this->source, $lineStartPos, $newOffset - $lineStartPos ) );
 		} else {
-			$this->curSourceColumn += self::getJsLength(
+			$this->curSourceColumn += Utils::getJsLength(
 				substr( $this->source, $this->curSourceOffset, $length ) );
 		}
 		$this->curSourceOffset = $newOffset;
@@ -114,10 +114,10 @@ class MappingsGenerator {
 		if ( $lineCount ) {
 			$lineStartPos = strrpos( $out, "\n" ) + 1;
 			$this->curOutLine += $lineCount;
-			$this->curOutColumn = self::getJsLength(
+			$this->curOutColumn = Utils::getJsLength(
 				substr( $out, $lineStartPos, strlen( $out ) - $lineStartPos ) );
 		} else {
-			$this->curOutColumn += self::getJsLength( $out );
+			$this->curOutColumn += Utils::getJsLength( $out );
 		}
 	}
 
@@ -151,7 +151,7 @@ class MappingsGenerator {
 		$this->prevSourceLine = $this->curSourceLine;
 		$this->prevSourceColumn = $this->curSourceColumn;
 
-		$this->curOutColumn += self::getJsLength( $out );
+		$this->curOutColumn += Utils::getJsLength( $out );
 	}
 
 	/**
@@ -175,16 +175,6 @@ class MappingsGenerator {
 		} while ( $vlq > 0 );
 
 		$this->mappings .= $encoded;
-	}
-
-	/**
-	 * Get the length of a string in UTF-16 code units
-	 *
-	 * @param string $s
-	 * @return int
-	 */
-	private static function getJsLength( $s ) {
-		return strlen( mb_convert_encoding( $s, 'UTF-16', 'UTF-8' ) ) / 2;
 	}
 
 	/**